doc/reference/index: Remove "Now that you have an overview" paragraph
This wording was originally from c2760945 (At the end, point to docs
that might be interesting next, 2016-05-17, #2688), which added it to
the end of the guide (where telling readers what they know makes some
sense).  It was moved to a "Cargo in Depth" section with db5be159
([src/doc/book] Move a paragraph to cargo-in-depth.md, 2017-08-31,
#4453), where it makes a bit less sense.  When that section became the
reference index in ca690628 ([doc/book] Create dir for book sections,
2017-08-31, #4455) the context assumed by the paragraph was completely
missing.

This commit removes the paragraph, which doesn't reduce the usefulness
of the reference index.  And the removal avoids confusing readers who
start with the reference docs and may now have the assumed overview.